Change host_id to be uint8_t

+++ template_cpp/src/include/ack_set.hpp
@@ -10,7 +10,7 @@ class AckSet
 {
 private:
     std::mutex lock;
-    std::unordered_map<uint32_t, std::unordered_set<uint64_t>> acked_messages;
+    std::unordered_map<uint32_t, std::unordered_set<uint8_t>> acked_messages;
     HostLookup host_lookup;
 
 public:
@@ -18,11 +18,11 @@ public:
 
     bool is_acked(TransportMessage message)
     {
-        uint64_t host_id = this->host_lookup.get_host_id_by_ip(message.address);
+        uint8_t host_id = this->host_lookup.get_host_id_by_ip(message.address);
         this->lock.lock();
         auto iterator = this->acked_messages.find(message.get_id());
         auto set =
-            iterator == this->acked_messages.cend() ? std::unordered_set<uint64_t>() : iterator->second;
+            iterator == this->acked_messages.cend() ? std::unordered_set<uint8_t>() : iterator->second;
         bool found = !(set.find(host_id) == set.cend());
         this->lock.unlock();
 
@@ -31,12 +31,12 @@ public:
 
     void ack(TransportMessage message)
     {
-        uint64_t host_id = this->host_lookup.get_host_id_by_ip(message.address);
+        uint8_t host_id = this->host_lookup.get_host_id_by_ip(message.address);
 
         this->lock.lock();
         if (this->acked_messages.find(message.get_id()) == this->acked_messages.cend())
         {
-            std::unordered_set<uint64_t> set;
+            std::unordered_set<uint8_t> set;
             set.insert(host_id);
             this->acked_messages.emplace(message.get_id(), set);
             this->lock.unlock();
