+++ template_cpp/src/include/host_lookup.hpp
@@ -10,8 +10,8 @@
 class HostLookup
 {
 private:
-    std::unordered_map<uint32_t, std::unordered_map<uint16_t, uint64_t>> address_to_host;
-    std::unordered_map<uint64_t, Address> host_to_address;
+    std::unordered_map<uint32_t, std::unordered_map<uint16_t, uint8_t>> address_to_host;
+    std::unordered_map<uint8_t, Address> host_to_address;
 
 public:
     HostLookup(std::string file_name)
@@ -27,7 +27,7 @@ public:
             // splitted[2] is the port
             std::vector<std::string> split_vec = StringHelpers::split(line, " ");
 
-            uint64_t id = std::stoul(split_vec[0]);
+            uint8_t id = static_cast<uint8_t>(std::stoul(split_vec[0]));
             std::string ip = split_vec[1] == "localhost" ? "127.0.0.1" : split_vec[1];
             Address address = Address(ip, static_cast<uint16_t>(std::stoul(split_vec[2])));
             host_to_address.emplace(id, address);
@@ -38,7 +38,7 @@ public:
     // In a system of n identifiers, the processes have ids from 1 to n
     // this function can therefore return 0 if it cannot find a process
     // with specified ip.
-    uint64_t get_host_id_by_ip(Address address) const
+    uint8_t get_host_id_by_ip(Address address) const
     {
         auto inner_map = address_to_host.find(address.ip);
         if (inner_map == address_to_host.end())
@@ -55,16 +55,16 @@ public:
         return id->second;
     }
 
-    Address get_address_by_host_id(uint64_t id) const
+    Address get_address_by_host_id(uint8_t id) const
     {
         auto pair = host_to_address.find(id);
 
         return pair == host_to_address.end() ? Address(0, 0) : pair->second;
     }
 
-    std::vector<uint64_t> get_hosts()
+    std::vector<uint8_t> get_hosts()
     {
-        std::vector<uint64_t> hosts(this->host_to_address.size());
+        std::vector<uint8_t> hosts(this->host_to_address.size());
         for (auto pair : this->host_to_address)
         {
             hosts.push_back(pair.first);
